requirements.txt
psycopg2
fastapi
uvicorn
mysql-connector-python